azure_functions/requirements.txt
azure-functions

# Project dependencies (kept in sync with repo root requirements.txt)
requests>=2.28.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
openpyxl>=3.1.0
xgboost>=2.0.0
flask>=3.0.0
azure-storage-blob>=12.0.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
python-dotenv>=1.0.0
pyarrow>=14.0.0